fix(soundboard): stop bundling mongoose/@sentry/node into the client build
Task 7 shared `requireActor` between audio-server-fns.ts and
soundboard-server-fns.ts via a static export/import pair. TanStack Start
strips `.handler()` bodies from the client bundle, so `requireActor` was
previously unreachable dead code there — but a static import edge outside
any handler body made the whole chunk (and its `~/server/session` ->
mongoose/@sentry/node chain) reachable from the client module graph again,
breaking `npm run build` ("subscribe" is not exported by
"__vite-browser-external", from @sentry/node-core's httpServerIntegration).

Move requireActor into its own module (app/utils/require-actor.ts) and
reach it via a dynamic `await import(...)` inside every handler in both
wrapper files, never a module-scope import/export. Same single definition,
no duplication — only the import mechanism changed.
