+++ app/utils/audio-server-fns.ts
@@ -24,16 +24,24 @@ import {
 // and the HeadObject size check in confirmAudioUpload can't drift between
 // the two callers.
 //
-// `~/server/session` is imported dynamically, inside each handler, rather
-// than at module scope: a static `import ... from '~/server/session'` would
-// (a) trip the `no-restricted-imports` lint rule that keeps ~/server/* out
-// of app/utils/**, and (b) pull server-only code (jose, cookie helpers) into
-// the client bundle. `createServerFn`'s `.handler()` body only ever executes
-// on the server — the client bundle gets a stub that calls the network
-// endpoint instead — so a dynamic import inside it never reaches the
+// `requireActor` (identity resolution — see ~/utils/require-actor.ts for the
+// full rationale) is reached via a dynamic `await import('~/utils/require-
+// actor')` INSIDE each handler below, never a module-scope `import`, and it
+// is never re-exported from this module either. A static import/export pair
+// used to exist here (this file defined `requireActor` and exported it for
+// `~/utils/soundboard-server-fns.ts` to import) and broke `npm run build`:
+// see ~/utils/require-actor.ts's doc comment for the exact mechanism. The
+// short version: a static import edge into (or out of) a module makes the
+// imported code reachable from the client's module graph even when every
+// call site is inside a `.handler()` body that TanStack Start strips from
+// the client bundle — the bundler can't know "only ever called from inside
+// a handler" from the import graph alone. Dynamic imports, scoped inside
+// each handler, avoid that: `createServerFn`'s `.handler()` body only ever
+// executes on the server, so the dynamic import inside it never reaches the
 // browser. This mirrors the existing pattern in ~/utils/uploadToR2.ts and
 // ~/server/functions/rpc.ts, which dynamically import their server-only
-// implementation modules the same way.
+// implementation modules the same way — and now applies to `requireActor`
+// itself, not just `~/server/session` underneath it.
 //
 // `SessionUser.id` is the OAuth provider's subject id (see
 // `toSessionUser`/`upsertUser` in `~/server/utils/oauth.ts`), not this app's
@@ -47,35 +55,20 @@ import {
 // every call, for every user — caught by this task's E2E suite hitting a
 // genuinely seeded, real `ownerId`.
 //
-// It returns BOTH ids, because the two are used for different things and
-// conflating them is what produced the split telemetry identity this branch
-// shipped with. `userId` (Mongo `_id`) is the only value that may scope a
-// query; `sessionUserId` (the OAuth provider id) is the identity every other
-// server function in this codebase tags telemetry with, so the same human
-// stays one person in GlitchTip and Umami whether they are uploading an image
-// or an audio file. See the `Actor` type in `~/server/functions/audio.ts`.
-//
-// Exported (not module-private) so `~/utils/soundboard-server-fns.ts` can
-// import and reuse it rather than defining a second copy — every wrapper in
-// this codebase that resolves the browser session to a Mongo `userId` needs
-// the exact same provider-id-to-`_id` resolution, and two copies is how they
-// drift (see this branch's phase-1 postmortem, above).
-export async function requireActor(): Promise<{ userId: string; sessionUserId: string }> {
-  const { getSession } = await import('~/server/session');
-  const { connectDB } = await import('~/server/db/connection');
-  const { User } = await import('~/server/db/models/User');
-  const session = await getSession();
-  if (!session) throw new Error('Not authenticated');
-  await connectDB();
-  const dbUser = await User.findOne({ providerId: session.id }).select('_id').lean();
-  if (!dbUser) throw new Error('User not found');
-  return { userId: String(dbUser._id), sessionUserId: session.id };
-}
+// `requireActor()` returns BOTH ids, because the two are used for different
+// things and conflating them is what produced the split telemetry identity
+// this branch shipped with. `userId` (Mongo `_id`) is the only value that
+// may scope a query; `sessionUserId` (the OAuth provider id) is the identity
+// every other server function in this codebase tags telemetry with, so the
+// same human stays one person in GlitchTip and Umami whether they are
+// uploading an image or an audio file. See the `Actor` type in
+// `~/server/functions/audio.ts`.
 
 export const createAudioUploadFn = createServerFn({ method: 'POST' })
   .inputValidator(createAudioUploadSchema)
   .handler(async ({ data }) => {
     const { createAudioUpload } = await import('~/server/functions/audio');
+    const { requireActor } = await import('~/utils/require-actor');
     return createAudioUpload({ data, ...(await requireActor()) });
   });
 
@@ -83,6 +76,7 @@ export const confirmAudioUploadFn = createServerFn({ method: 'POST' })
   .inputValidator(confirmAudioUploadSchema)
   .handler(async ({ data }) => {
     const { confirmAudioUpload } = await import('~/server/functions/audio');
+    const { requireActor } = await import('~/utils/require-actor');
     return confirmAudioUpload({ data, ...(await requireActor()) });
   });
 
@@ -90,6 +84,7 @@ export const listAudioAssetsFn = createServerFn({ method: 'POST' })
   .inputValidator(listAudioAssetsSchema)
   .handler(async ({ data }) => {
     const { listAudioAssets } = await import('~/server/functions/audio');
+    const { requireActor } = await import('~/utils/require-actor');
     return listAudioAssets({ data, ...(await requireActor()) });
   });
 
@@ -97,6 +92,7 @@ export const updateAudioAssetFn = createServerFn({ method: 'POST' })
   .inputValidator(updateAudioAssetSchema)
   .handler(async ({ data }) => {
     const { updateAudioAsset } = await import('~/server/functions/audio');
+    const { requireActor } = await import('~/utils/require-actor');
     return updateAudioAsset({ data, ...(await requireActor()) });
   });
 
@@ -104,6 +100,7 @@ export const bulkTagAudioAssetsFn = createServerFn({ method: 'POST' })
   .inputValidator(bulkTagAudioAssetsSchema)
   .handler(async ({ data }) => {
     const { bulkTagAudioAssets } = await import('~/server/functions/audio');
+    const { requireActor } = await import('~/utils/require-actor');
     return bulkTagAudioAssets({ data, ...(await requireActor()) });
   });
 
@@ -111,6 +108,7 @@ export const retryAudioAssetFn = createServerFn({ method: 'POST' })
   .inputValidator(retryAudioAssetSchema)
   .handler(async ({ data }) => {
     const { retryAudioAsset } = await import('~/server/functions/audio');
+    const { requireActor } = await import('~/utils/require-actor');
     return retryAudioAsset({ data, ...(await requireActor()) });
   });
 
@@ -118,6 +116,7 @@ export const deleteAudioAssetFn = createServerFn({ method: 'POST' })
   .inputValidator(deleteAudioAssetSchema)
   .handler(async ({ data }) => {
     const { deleteAudioAsset } = await import('~/server/functions/audio');
+    const { requireActor } = await import('~/utils/require-actor');
     return deleteAudioAsset({ data, ...(await requireActor()) });
   });
 
@@ -132,6 +131,7 @@ export const scanOrphanAudioFn = createServerFn({ method: 'POST' })
   .inputValidator(scanOrphanAudioSchema)
   .handler(async ({ data }) => {
     const { scanOrphanAudio } = await import('~/server/functions/audio-cleanup');
+    const { requireActor } = await import('~/utils/require-actor');
     return scanOrphanAudio({ data, ...(await requireActor()) });
   });
 
@@ -139,5 +139,6 @@ export const deleteOrphanAudioFn = createServerFn({ method: 'POST' })
   .inputValidator(deleteOrphanAudioSchema)
   .handler(async ({ data }) => {
     const { deleteOrphanAudio } = await import('~/server/functions/audio-cleanup');
+    const { requireActor } = await import('~/utils/require-actor');
     return deleteOrphanAudio({ data, ...(await requireActor()) });
   });
