+++ app/utils/soundboard-server-fns.ts
@@ -9,22 +9,25 @@ import {
   loadBoardStateSchema,
   saveBoardStateSchema,
 } from '~/types/schemas/soundboard';
-import { requireActor } from '~/utils/audio-server-fns';
 
 // ---------------------------------------------------------------------------
 // Browser-facing server-fn wrappers for ~/server/functions/packages and
 // ~/server/functions/soundboard.
 //
 // Structure copied exactly from ~/utils/audio-server-fns.ts: `requireActor`
-// (imported from there, not re-defined — see its doc comment) resolves the
-// session's OAuth provider id to this app's Mongo `_id` before anything below
-// touches a query, and `~/server/session` (transitively, inside
-// `requireActor`) is only ever reached via a dynamic import, never a
-// module-scope one — a static import would (a) trip the `no-restricted-
-// imports` lint rule that keeps ~/server/* out of app/utils/**, and (b) pull
-// server-only code into the client bundle. `createServerFn`'s `.handler()`
-// body only ever executes on the server, so the dynamic imports inside each
-// handler below never reach the browser.
+// (see ~/utils/require-actor.ts) resolves the session's OAuth provider id to
+// this app's Mongo `_id` before anything below touches a query.
+//
+// `requireActor` is reached via a dynamic `await import('~/utils/require-
+// actor')` INSIDE each handler below, never a module-scope `import` — this
+// file previously had `import { requireActor } from '~/utils/audio-server-
+// fns'` at module scope, which broke `npm run build` (see ~/utils/require-
+// actor.ts's doc comment for the exact mechanism: a static import edge makes
+// the imported module's `~/server/session` chain reachable from the client's
+// module graph even though every call site is inside a stripped `.handler()`
+// body). `createServerFn`'s `.handler()` body only ever executes on the
+// server, so the dynamic imports inside each handler below never reach the
+// browser.
 //
 // `requireActor()` returns `{ userId, sessionUserId }`. `userId` (the Mongo
 // `_id`) is the only value that may scope a query — it is what every
@@ -48,6 +51,7 @@ import { requireActor } from '~/utils/audio-server-fns';
 
 export const listPackagesFn = createServerFn({ method: 'GET' }).handler(async () => {
   const { listPackages } = await import('~/server/functions/packages');
+  const { requireActor } = await import('~/utils/require-actor');
   return listPackages(await requireActor());
 });
 
@@ -55,6 +59,7 @@ export const getPackageFn = createServerFn({ method: 'GET' })
   .inputValidator(getPackageSchema)
   .handler(async ({ data }) => {
     const { getPackage } = await import('~/server/functions/packages');
+    const { requireActor } = await import('~/utils/require-actor');
     return getPackage({ data, ...(await requireActor()) });
   });
 
@@ -62,6 +67,7 @@ export const createPackageFn = createServerFn({ method: 'POST' })
   .inputValidator(createPackageSchema)
   .handler(async ({ data }) => {
     const { createPackage } = await import('~/server/functions/packages');
+    const { requireActor } = await import('~/utils/require-actor');
     return createPackage({ data, ...(await requireActor()) });
   });
 
@@ -69,6 +75,7 @@ export const updatePackageFn = createServerFn({ method: 'POST' })
   .inputValidator(updatePackageSchema)
   .handler(async ({ data }) => {
     const { updatePackage } = await import('~/server/functions/packages');
+    const { requireActor } = await import('~/utils/require-actor');
     return updatePackage({ data, ...(await requireActor()) });
   });
 
@@ -76,6 +83,7 @@ export const deletePackageFn = createServerFn({ method: 'POST' })
   .inputValidator(deletePackageSchema)
   .handler(async ({ data }) => {
     const { deletePackage } = await import('~/server/functions/packages');
+    const { requireActor } = await import('~/utils/require-actor');
     return deletePackage({ data, ...(await requireActor()) });
   });
 
@@ -83,6 +91,7 @@ export const clonePackageFn = createServerFn({ method: 'POST' })
   .inputValidator(clonePackageSchema)
   .handler(async ({ data }) => {
     const { clonePackage } = await import('~/server/functions/packages');
+    const { requireActor } = await import('~/utils/require-actor');
     return clonePackage({ data, ...(await requireActor()) });
   });
 
@@ -94,6 +103,7 @@ export const listPackageAssetsFn = createServerFn({ method: 'GET' })
   .inputValidator(listPackageAssetsSchema)
   .handler(async ({ data }) => {
     const { listPackageAssets } = await import('~/server/functions/packages');
+    const { requireActor } = await import('~/utils/require-actor');
     return listPackageAssets({ data, ...(await requireActor()) });
   });
 
@@ -101,6 +111,7 @@ export const loadBoardStateFn = createServerFn({ method: 'GET' })
   .inputValidator(loadBoardStateSchema)
   .handler(async ({ data }) => {
     const { loadBoardState } = await import('~/server/functions/soundboard');
+    const { requireActor } = await import('~/utils/require-actor');
     return loadBoardState({ data, ...(await requireActor()) });
   });
 
@@ -108,5 +119,6 @@ export const saveBoardStateFn = createServerFn({ method: 'POST' })
   .inputValidator(saveBoardStateSchema)
   .handler(async ({ data }) => {
     const { saveBoardState } = await import('~/server/functions/soundboard');
+    const { requireActor } = await import('~/utils/require-actor');
     return saveBoardState({ data, ...(await requireActor()) });
   });
